@@ -5,11 +5,11 @@
   "description": "My Electron application description",
   "main": ".webpack/main",
   "scripts": {
-    "start": "electron-forge start",
+    "start": "npm run lint && electron-forge start",
     "package": "electron-forge package",
     "make": "electron-forge make",
     "publish": "electron-forge publish",
-    "lint": "eslint --ext .ts,.tsx ."
+    "lint": "eslint ."
   },
   "keywords": [],
   "author": "kinator",
@@ -24,31 +24,36 @@
     "@electron-forge/plugin-fuses": "^7.8.1",
     "@electron-forge/plugin-webpack": "^7.8.1",
     "@electron/fuses": "^1.8.0",
+    "@eslint/compat": "^1.3.2",
+    "@eslint/js": "^9.33.0",
     "@tailwindcss/postcss": "^4.1.11",
+    "@types/electron-squirrel-startup": "^1.0.2",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
     "@vercel/webpack-asset-relocator-loader": "^1.7.3",
     "autoprefixer": "^10.4.21",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "css-loader": "^6.11.0",
     "electron": "37.2.1",
-    "eslint": "^8.57.1",
-    "eslint-plugin-import": "^2.32.0",
+    "eslint": "^9.33.0",
+    "eslint-plugin-react": "^7.37.5",
     "fork-ts-checker-webpack-plugin": "^7.3.0",
+    "globals": "^16.3.0",
+    "jiti": "^2.5.1",
     "lucide-react": "^0.525.0",
     "node-loader": "^2.1.0",
     "postcss": "^8.5.6",
     "postcss-loader": "^8.1.1",
+    "prettier": "^3.6.2",
     "style-loader": "^3.3.4",
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^4.1.11",
     "ts-loader": "^9.5.2",
     "ts-node": "^10.9.2",
     "tw-animate-css": "^1.3.6",
-    "typescript": "~4.5.4"
+    "typescript": "^5.9.2",
+    "typescript-eslint": "^8.39.1"
   },
   "dependencies": {
     "@radix-ui/react-accordion": "^1.2.11",
