@@ -1,5 +1,5 @@
-const { contextBridge, ipcRenderer } = require('electron/renderer')
+import { contextBridge, ipcRenderer } from "electron/renderer";
 
-contextBridge.exposeInMainWorld('electronAPI', {
-  saveMasterPassword: () => ipcRenderer.invoke('masterPassword:save')
-})
+contextBridge.exposeInMainWorld("electronAPI", {
+  saveMasterPassword: () => ipcRenderer.invoke("masterPassword:save"),
+});
