@@ -1,5 +1,5 @@
 export default {
-    plugins: {
-      "@tailwindcss/postcss": {},
-    }
-  }
+  plugins: {
+    "@tailwindcss/postcss": {},
+  },
+};
