@@ -1,17 +1,18 @@
-import { app, BrowserWindow, ipcMain } from 'electron';
+import { app, BrowserWindow, ipcMain } from "electron";
+import electronSquirrelStartup from "electron-squirrel-startup";
 // import * as passwordManager from './managePassword';
 declare const MAIN_WINDOW_WEBPACK_ENTRY: string;
 declare const MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY: string;
 
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
-if (require('electron-squirrel-startup')) {
+if (electronSquirrelStartup) {
   app.quit();
 }
 
-function saveMasterPassword(){
+function saveMasterPassword() {
   // passwordManager.savePassword(masterPassword)
-  console.log('SAVING MASTER PASSWORD')
-  return 'SAVED TO KEYTAR'
+  console.log("SAVING MASTER PASSWORD");
+  return "SAVED TO KEYTAR";
 }
 
 const createWindow = (): void => {
@@ -32,18 +33,18 @@ const createWindow = (): void => {
 };
 
 app.whenReady().then(() => {
-  ipcMain.handle('masterPassword:save', saveMasterPassword)
-  createWindow()
-  app.on('activate', function () {
-    if (BrowserWindow.getAllWindows().length === 0) createWindow()
-  })
-})
+  ipcMain.handle("masterPassword:save", saveMasterPassword);
+  createWindow();
+  app.on("activate", function () {
+    if (BrowserWindow.getAllWindows().length === 0) createWindow();
+  });
+});
 
 // Quit when all windows are closed, except on macOS. There, it's common
 // for applications and their menu bar to stay active until the user quits
 // explicitly with Cmd + Q.
-app.on('window-all-closed', () => {
-  if (process.platform !== 'darwin') {
+app.on("window-all-closed", () => {
+  if (process.platform !== "darwin") {
     app.quit();
   }
 });
