@@ -1,7 +1,4 @@
 import LITERALS from "./literals";
 import REG_EX from "./regEx";
 
-export {
-    LITERALS,
-    REG_EX
-}
+export { LITERALS, REG_EX };
