@@ -1,9 +1,10 @@
-const LITERALS =  {
-MIN_PASSWORD_ERROR_MESSAGE: 'Password must be at least 8 characters long',
-PASSWORD_CONTAINS_LOWER_CHAR :"Password must contain at least one lowercase letter",
-PASSWORD_CONTAINS_UPPER_CHAR: "Password must contain at least one uppercase letter",
-PASSWORD_CONTAINS_NUMBER: "Password must contain at least one number",
-PASSWORD_CONTAINS_SPECIAL_CHAR: "Password must contain at least one special character"
-}
+const LITERALS = {
+  MIN_PASSWORD_ERROR_MESSAGE: "Password must be at least 8 characters long",
+  PASSWORD_CONTAINS_LOWER_CHAR: "Password must contain at least one lowercase letter",
+  PASSWORD_CONTAINS_UPPER_CHAR: "Password must contain at least one uppercase letter",
+  PASSWORD_CONTAINS_NUMBER: "Password must contain at least one number",
+  PASSWORD_CONTAINS_SPECIAL_CHAR:
+    "Password must contain at least one special character",
+};
 
 export default LITERALS;
