@@ -1,8 +1,8 @@
 const REG_EX = {
-    atLeastOneLowerChar: /[a-z]/,
-    atLeastOneUpperChar: /[A-Z]/,
-    atLeastOneNumber: /[0-9]/,
-    atLeastOneSpecialChar: /[!@#$%^&*()_\-+=\[\]{};':"\\|,.<>\/?`~]/,
-  };
-  
+  atLeastOneLowerChar: /[a-z]/,
+  atLeastOneUpperChar: /[A-Z]/,
+  atLeastOneNumber: /[0-9]/,
+  atLeastOneSpecialChar: /[!@#$%^&*()_\-+=[\]{};':"\\|,.<>/?`~]/,
+};
+
 export default REG_EX;
