@@ -1,7 +1,5 @@
-function Dashboard(){
-    return (
-        <p>This is Dashboard Component !!</p>
-    )
+function Dashboard() {
+  return <p>This is Dashboard Component !!</p>;
 }
 
 export default Dashboard;
