@@ -1,10 +1,10 @@
-import type IForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
+import type IForkTsCheckerWebpackPlugin from "fork-ts-checker-webpack-plugin";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const ForkTsCheckerWebpackPlugin: typeof IForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
+const ForkTsCheckerWebpackPlugin: typeof IForkTsCheckerWebpackPlugin = require("fork-ts-checker-webpack-plugin");
 
 export const plugins = [
   new ForkTsCheckerWebpackPlugin({
-    logger: 'webpack-infrastructure',
+    logger: "webpack-infrastructure",
   }),
 ];
