@@ -1,15 +1,15 @@
-import type { Configuration } from 'webpack';
-import path from 'path';
+import type { Configuration } from "webpack";
+import path from "path";
 
-import { rules } from './webpack.rules';
-import { plugins } from './webpack.plugins';
+import { rules } from "./webpack.rules";
+import { plugins } from "./webpack.plugins";
 
 rules.push({
   test: /\.css$/,
   use: [
-    { loader: 'style-loader' },
-    { loader: 'css-loader' },
-    { loader: 'postcss-loader' },
+    { loader: "style-loader" },
+    { loader: "css-loader" },
+    { loader: "postcss-loader" },
   ],
 });
 
@@ -19,9 +19,9 @@ export const rendererConfig: Configuration = {
   },
   plugins,
   resolve: {
-    extensions: ['.js', '.ts', '.jsx', '.tsx', '.css'],
+    extensions: [".js", ".ts", ".jsx", ".tsx", ".css"],
     alias: {
-      '@': path.resolve(__dirname, 'src/renderer'),
+      "@": path.resolve(__dirname, "src/renderer"),
     },
   },
 };
