@@ -1,20 +1,20 @@
-import type { ModuleOptions } from 'webpack';
+import type { ModuleOptions } from "webpack";
 
-export const rules: Required<ModuleOptions>['rules'] = [
+export const rules: Required<ModuleOptions>["rules"] = [
   // Add support for native node modules
   {
     // We're specifying native_modules in the test because the asset relocator loader generates a
     // "fake" .node file which is really a cjs file.
     test: /native_modules[/\\].+\.node$/,
-    use: 'node-loader',
+    use: "node-loader",
   },
   {
     test: /[/\\]node_modules[/\\].+\.(m?js|node)$/,
     parser: { amd: false },
     use: {
-      loader: '@vercel/webpack-asset-relocator-loader',
+      loader: "@vercel/webpack-asset-relocator-loader",
       options: {
-        outputAssetBase: 'native_modules',
+        outputAssetBase: "native_modules",
       },
     },
   },
@@ -22,7 +22,7 @@ export const rules: Required<ModuleOptions>['rules'] = [
     test: /\.tsx?$/,
     exclude: /(node_modules|\.webpack)/,
     use: {
-      loader: 'ts-loader',
+      loader: "ts-loader",
       options: {
         transpileOnly: true,
       },
